requirements.txt
py-cord[voice] >= 2.0.0b5
python-mpd2 >= 3.0.5